context_agent/requirements.txt
fastapi
uvicorn
kafka-python
redis
pydantic